Update routes sopDetail -> category
